feat: Run garbage-collect before each sample in basic-stats-benchmark
This is in a hope to reduce GC runs that may be caused by a previous
sample. For example, in a scenario where sample-1 is run before sample-2, and
sample-1 uses a lot of memory (bot just shy from triggerring GC) then GC may be
triggered when running sample-2, although the latter may not use enough memory
to trigger GC by itself (nor be a close to it).